# docs/swagger.yaml
basePath: /
definitions:
...
  models.Response:
    properties:
      message:
        type: string
    type: object
host: localhost:3002
info:
  contact: {}
  description: This is a simple API service for Dog Breeds consume from the Dog API.
  title: API for Dog Breeds
  version: "1.0"
paths:
...
  /dogs/{id}:
    delete:
      consumes:
      - application/json
      description: delete a dog
      operationId: delete-dog
      parameters:
      - description: Dog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a dog
    put:
      consumes:
      - application/json
      description: update dog by ID
      operationId: update-dog-by-id
      parameters:
      - description: Dog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update dog
swagger: "2.0"